api/requirements.txt
fastapi
uvicorn
requests
sqlmodel
python-multipart
python-jose[cryptography]
passlib[argon2]
